Move Strategies helper to utilities
